Refactoring
- remove redundant methods/unit tests
- add loggers to services
- add null checks for services
- rewrite methods

+++ src/main/java/ru/alikhano/cyberlife/service/impl/CartServiceImpl.java
@@ -4,9 +4,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
+
 import ru.alikhano.cyberlife.dto.CartDTO;
 
 import ru.alikhano.cyberlife.dao.CartDao;
@@ -17,6 +21,8 @@ import ru.alikhano.cyberlife.service.CartService;
 
 @Service
 public class CartServiceImpl implements CartService {
+
+	private static final Logger LOGGER = LogManager.getLogger(CartServiceImpl.class);
 	
 	@Autowired
 	private CartDao cartDao;
@@ -57,9 +63,7 @@ public class CartServiceImpl implements CartService {
 	@Override
 	@Transactional
 	public void update(CartDTO cartDTO) {
-
 		merge(cartDTO);
-		
 	}
 
 	/**
@@ -68,9 +72,7 @@ public class CartServiceImpl implements CartService {
 	@Override
 	@Transactional
 	public void create(CartDTO cartDTO) {
-
 		cartDao.create(cartMapper.backward(cartDTO));
-		
 	}
 
 	/**
@@ -79,7 +81,6 @@ public class CartServiceImpl implements CartService {
 	@Override
 	@Transactional
 	public int createAndGetId(CartDTO cartDTO) {
-
 		return cartDao.createAndGetId(cartMapper.backward(cartDTO));
 	}
 
@@ -88,23 +89,28 @@ public class CartServiceImpl implements CartService {
 	 */
 	@Override
 	public void merge(CartDTO cartDTO) {
-
 		cartDao.merge(cartMapper.backward(cartDTO));
 	}
 
 	@Override
 	public void deleteItemFromCart(CartDTO cartDTO, int itemId) {
+		if (cartDTO == null) {
+			LOGGER.info("Cart is null");
+			return;
+		}
 		Set<CartItemDTO> items = cartDTO.getItems();
+		if (CollectionUtils.isEmpty(items)) {
+			LOGGER.info("Empty cart - nothing to delete");
+			return;
+		}
 		for (CartItemDTO item : items) {
 			if (item.getItemId() == itemId) {
 				items.remove(item);
-				double grandTotal = cartDTO.getGrandTotal() - item.getTotalPrice();
 				cartItemService.delete(item);
-				cartDTO.setItems(items);
-				cartDTO.setGrandTotal(grandTotal);
+				cartDTO.setGrandTotal(cartDTO.getGrandTotal() - item.getTotalPrice());
 			}
 		}
-		if (items.isEmpty()) {
+		if (CollectionUtils.isEmpty(items)) {
 			cartDTO.setGrandTotal(0.0);
 		}
 		merge(cartDTO);
