+++ src/main/java/ru/alikhano/cyberlife/service/impl/OrderItemServiceImpl.java
@@ -1,8 +1,11 @@
 package ru.alikhano.cyberlife.service.impl;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 
 import ru.alikhano.cyberlife.dto.OrderDTO;
 import ru.alikhano.cyberlife.dto.OrderItemDTO;
@@ -13,6 +16,8 @@ import ru.alikhano.cyberlife.service.OrderItemService;
 
 @Service
 public class OrderItemServiceImpl implements OrderItemService {
+
+	private static final Logger LOGGER = LogManager.getLogger(OrderItemServiceImpl.class);
 	
 	@Autowired
 	private OrderItemDao orderItemDao;
@@ -26,6 +31,10 @@ public class OrderItemServiceImpl implements OrderItemService {
 	@Override
 	@Transactional
 	public void create(OrderItemDTO orderItemDTO) {
+		if (orderItemDTO == null) {
+			LOGGER.info("Order item is null");
+			return;
+		}
 		orderItemDao.create(orderItemMapper.backward(orderItemDTO));
 	}
 
@@ -35,6 +44,10 @@ public class OrderItemServiceImpl implements OrderItemService {
 	@Override
 	@Transactional
 	public void update(OrderItemDTO orderItemDTO) {
+		if (orderItemDTO == null) {
+			LOGGER.info("Order item is null");
+			return;
+		}
 		orderItemDao.update(orderItemMapper.backward(orderItemDTO));
 	}
 
@@ -44,6 +57,10 @@ public class OrderItemServiceImpl implements OrderItemService {
 	@Override
 	@Transactional
 	public void delete(OrderItemDTO orderItemDTO) {
+		if (orderItemDTO == null) {
+			LOGGER.info("Order item is null");
+			return;
+		}
 		orderItemDao.delete(orderItemMapper.backward(orderItemDTO));
 	}
 
@@ -53,6 +70,10 @@ public class OrderItemServiceImpl implements OrderItemService {
 	@Override
 	@Transactional
 	public void deleteAll(OrderDTO orderDTO) {
+		if (orderDTO == null || CollectionUtils.isEmpty(orderDTO.getOrderItems())) {
+			LOGGER.info("Order and/or order items are null");
+			return;
+		}
 		for (OrderItemDTO item : orderDTO.getOrderItems()) {
 			delete(item);
 		}
